+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog for Razorpay-PHP SDK. Follows [keepachangelog.com](http://keepachangel
 
 ## Unreleased
 ### Added
+- Support for VirtualAccount entity
 - Support for subscriptions
 
 ## [2.1.0] - 2017-10-10